requirements.txt
absl-py==1.3.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
audioread==3.0.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
chess==1.9.3
clang==5.0
colorama==0.4.5
contourpy==1.0.5
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
entrypoints==0.4
exceptiongroup==1.0.0rc9
executing==1.1.0
fastjsonschema==2.16.2
filelock==3.11.0
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.13.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==14.0.6
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
mistune==2.0.4
mpmath==1.3.0
msgpack==1.0.5
nest-asyncio==1.5.6
networkx==3.1
numba==0.56.4
oauthlib==3.2.0
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pickles==0.1.1
pickleshare==0.7.5

ply==3.11
pooch==1.6.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyee==8.2.2
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytube==12.1.2
pytz==2022.2.1
pywin32==304
pywinpty==2.0.8
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scrapingant-client==1.0.1
Send2Trash==1.8.0
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.5
sqlparse==0.4.2
stack-data==0.5.1
termcolor==1.1.0
terminado==0.16.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
tzdata==2022.2
urllib3==1.26.12
utm==0.7.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
Werkzeug==2.2.2
whitenoise==6.2.0
widgetsnbextension==4.0.3
windows-curses==2.3.0
wrapt==1.12.1
wsproto==1.2.0
zipp==3.8.1
